just map the data, don't create another router
