+++ pages/downloads/views/index.jsx
@@ -24,7 +24,7 @@ const Index = () => {
     <h3><Link page="downloads.report" report="ppl-list" label="Project list" /></h3>
     <p>Download a .csv file with summaries of the admin details of project licences such as expiry dates and statuses.</p>
 
-    <h3><Link page="downloads.pplConditionsReport" label="Project conditions and authorisations" /></h3>
+    <h3><Link page="downloads.report" report="ppl-conditions" label="Project conditions and authorisations" /></h3>
     <p>Download a .csv file with summaries of the conditions and authorisations for each project.</p>
 
     <h3><Link page="nts" label="Non-technical summaries by year" /></h3>
